add: ability to view specific user profiles depending on url params, output user data accoridng to url/id
