+++ client/src/index.css
@@ -71,6 +71,9 @@ th {
   .btn {
     @apply bg-grotto-100 font-semibold text-white p-2 rounded w-48 hover:bg-primary;
   }
+  .search-btn {
+    @apply bg-grotto-100 m-2 hover:bg-primary px-6 max-[399px]:px-2 py-1 text-xs lg:text-sm text-white rounded-full  outline outline-2 border-none outline-grotto-100;
+  }
   .modalCard {
     @apply flex flex-col bg-cardBg rounded-lg shadow-cardSh px-4 py-8 my-2 mx-auto w-[90%] max-w-[700px] relative;
   }
